adicionando exception

+++ src/main/java/com/mercadolibre/desafiotesting/service/LocalidadeService.java
@@ -2,9 +2,10 @@ package com.mercadolibre.desafiotesting.service;
 
 import com.mercadolibre.desafiotesting.model.Localidade;
 import com.mercadolibre.desafiotesting.repository.LocalidadeRepository;
-import org.apache.tomcat.jni.Local;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -21,6 +22,8 @@ public class LocalidadeService {
 
     public Localidade findById(Long id) {
         Localidade result = localidadeRepository.findById(id).orElse(null);
+        if (result == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Localidade with id: " + id + " not found");
         return result;
     }
 
@@ -36,10 +39,13 @@ public class LocalidadeService {
 
     public Localidade updateById(Long id, Localidade localidade) {
         Localidade novaLocalidade = localidadeRepository.getById(id);
+        if (novaLocalidade == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Localidade with id: " + id + " not found");
         novaLocalidade.setBairro(localidade.getBairro());
         novaLocalidade.setNome(localidade.getNome());
         novaLocalidade.setPreco(localidade.getPreco());
         Localidade result = localidadeRepository.save(novaLocalidade);
+
         return result;
     }
 
@@ -48,8 +54,10 @@ public class LocalidadeService {
     }
 
     public void deleteById(Long id) {
+        Localidade result = localidadeRepository.getById(id);
+        if (result == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Localidade with id: " + id + " not found");
         localidadeRepository.deleteById(id);
     }
-
 
 }
